requirements.txt
psycopg2-binary
pandas
fastapi
uvicorn[standard]
matplotlib
seaborn
folium
numpy
scikit-learn
xgboost
jupyter
python-dotenv
xmltodict
sqlalchemy